100% working, but mirrored
This exports fine, except that it exports as a mirrored image since Y axis on a CNC machine vs Processing is inverted.  Will require inverting all Y values on export (or in the drawing) but I don't have time for that right now.  You can make anything you want, just put the button stack/mcu pocket on the reverse side of where you actually want it.